Fixed error where .gitignore included Makefile.in files. Fixed error where an incorrect path was used in the elf32 Makefle.in file.
Made changes to command interpreter:
	* Replaced strcmpi calls in code to handle command-line arguments with single character comparisons
	* Fixed error where chdir() call for drive change command did not check if it returned -1
	* Cleaned up code in critical_error_handler()

Added function get_kernel_stack_base(), set_kernel_stack_base().
Added kernelstackbase member to PROCESS typedef struct.
Added function realloc_kernel(), realloc_user().
Reordered dispatch handler code.
Added function GetVersion().
Replace define CCPVER with CCP_MAJOR_VERSION, CCP_MINOR_VERSION, CCP_RELEASE_VERSION.
Updated dispatch_handler() to call GetVersion().

+++ modules/binfmt/elf32/elf32.c
@@ -12,7 +12,7 @@
 
 #define MODULE_INIT elf32_init
 
-int elf32_init(char *init) {
+size_t elf32_init(char *init) {
 EXECUTABLEFORMAT exec;
 uint8_t magicnumber[] = { 0x7f,'E','L','F',1,1,1 };
